do not set pod cpu limits by default
Pod CPU limits are generally a bad thing. When they are in place, k8s
attempts to throttle the pod to that limit. This can typically result in
very poor utilisation of the node.

Take for example an executor pod that is configured with a hard limit of
2 cpus. It could be scheduled onto a node with the equivalent of 16
CPUs. However, due to the limit, it will only ever used the equivalent
of 2 CPUs, leaving the bulk of the machine unutilised.

We have also experienced issues where CPU limits cause instability or
cause processes to crash.