use relative imports and a static template load
Using imports that are hard-coded to the module name makes this lib
unable to work when being directly loaded as a submodule.

Relying on the built-in Airflow template search behaviour also causes
problems, because the location where the library lives is unlikely to be
included in the Airflow template search path. We can resolve this by
jsut directly loading the template from its relative position to the
__file__ magic var, and passing the resulting string into the
CustomizableSparkKubernetesOperator to be templated. This works because
we do the additional round of templating later at task execution time.